+++ docs/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog for Typopo
 
 ## 2.2.2 // 2019-10-
-* fix dashes between dates, e.g. 2019 - 08-02 → 2019–08–02 (places ndash)
+* Fix dashes between dates, e.g. 2019 - 08-02 → 2019–08–02 (places ndash)
+* Fix dashes between percentage range, e.g. 20% - 30% → 20%–30%
 
 ## 2.2.1 // 2019-03-02
 ### New features